chore: split out kaioken signal into files
