+++ packages/lib/src/reconciler.ts
@@ -2,7 +2,7 @@ import type { AppContext } from "./appContext"
 import { ELEMENT_TYPE, FLAG, $FRAGMENT } from "./constants.js"
 import { ctx } from "./globals.js"
 import { isVNode } from "./utils.js"
-import { Signal } from "./signal.js"
+import { Signal } from "./signals"
 import { __DEV__ } from "./env.js"
 import { createElement, Fragment } from "./element.js"
 import { bitmapOps } from "./bitmap.js"
