+++ packages/lib/src/ssr/server.ts
@@ -8,7 +8,7 @@ import {
   propsToElementAttributes,
   selfClosingTags,
 } from "../utils.js"
-import { Signal } from "../signal.js"
+import { Signal } from "../signals"
 import { $CONTEXT_PROVIDER, ELEMENT_TYPE, $FRAGMENT } from "../constants.js"
 import { assertValidElementProps } from "../props.js"
