+++ packages/lib/tsconfig.json
@@ -24,6 +24,12 @@
     "jsxFactory": "kaioken.createElement",
     "jsxFragmentFactory": "kaioken.Fragment"
   },
-  "include": ["./src/*", "./src/tests", "src/types.ts", "./src/ssr/*"],
+  "include": [
+    "./src/*",
+    "./src/tests",
+    "./src/signals",
+    "src/types.ts",
+    "./src/ssr/*"
+  ],
   "exclude": ["dist"]
 }
