+++ sandbox/csr/src/components/SignalsExample.tsx
@@ -4,11 +4,21 @@ const count = signal(0, "count")
 const isTracking = signal(false, "isTracking")
 const double = computed(() => {
   if (isTracking.value) {
-    return count.value * 100
+    console.log("hello world 123", count.value)
+    return count.value * 2
   }
 
   return 0
 }, "double")
+
+const quadruple = computed(() => {
+  console.log("quad 123", count.value)
+  return count.value * 2
+}, "double")
+
+const watcher = watch(() => {
+  console.log("count 123", count.value)
+})
 
 export function SignalsExample() {
   return (
@@ -36,10 +46,6 @@ const GlobalComputedExample = () => {
     count.value += 1
   }
 
-  const watcher = watch(() => {
-    console.log("count 123", count.value)
-  })
-
   const onSwitch = () => {
     isTracking.value = !isTracking.value
     console.log("calling on switch method")
@@ -47,7 +53,7 @@ const GlobalComputedExample = () => {
 
   return (
     <div ref={refTest} className="flex flex-col">
-      <h1>count: {count.value}</h1>
+      <h1>count: {count}</h1>
       <h1>Double: {double}</h1>
       <h1>is tracking: {`${isTracking}`}</h1>
